chat: refactor NewDm in preparation for new flow
Extract reusable logic that will be used in other views into a new hook.
